Fix shake-feed ingest failing on untyped LEAST parameters
Postgres cannot infer parameter types used only inside LEAST($4, $5)
in a VALUES list ('could not determine data type of parameter'), so
every ingest request 500'd and the task never counted. Cast both to
int. Also stop discarding shakes beyond the per-report clamp on flush —
the remainder rides along with the next batch.

+++ src/services/marketService.js
@@ -3048,12 +3048,14 @@ export async function ingestShakeFeed(input) {
     username: input.username,
     first_name: input.first_name,
   });
+  // Касты обязательны: параметры только внутри LEAST не дают Postgres
+  // вывести тип ("could not determine data type of parameter").
   await query(
     `
       INSERT INTO user_task_events (user_id, day_key, event_key, count)
-      VALUES ($1, $2, $3, LEAST($4, $5))
+      VALUES ($1, $2, $3, LEAST($4::int, $5::int))
       ON CONFLICT (user_id, day_key, event_key)
-      DO UPDATE SET count = LEAST(user_task_events.count + $4, $5), updated_at = now()
+      DO UPDATE SET count = LEAST(user_task_events.count + $4::int, $5::int), updated_at = now()
     `,
     [user.id, getDayKey(), SHAKE_FEED_EVENT_KEY, count, SHAKE_FEED_DAILY_CAP],
   );
